requirements.txt
psycopg2
Flask 
Flask-SQLAlchemy
Flask-JWT-Extended
python-dotenv
Flask-Script
Flask-Migrate
passlib